# Faster Prepared Statement Execution by Using Raw SQL for Caching