--- a/go/cmd/vtgateclienttest/services/echo.go
+++ b/go/cmd/vtgateclienttest/services/echo.go
@@ -173,13 +173,13 @@ func (c *echoClient) VStream(ctx context.Context, tabletType topodatapb.TabletTy
 	return c.fallbackClient.VStream(ctx, tabletType, vgtid, filter, flags, callback)
 }
 
-func (c *echoClient) Prepare(ctx context.Context, session *vtgatepb.Session, sql string) (*vtgatepb.Session, []*querypb.Field, error) {
+func (c *echoClient) Prepare(ctx context.Context, session *vtgatepb.Session, sql string) (*vtgatepb.Session, []*querypb.Field, uint16, error) {
 	if strings.HasPrefix(sql, EchoPrefix) {
 		return session, echoQueryResult(map[string]any{
 			"callerId": callerid.EffectiveCallerIDFromContext(ctx),
 			"query":    sql,
 			"session":  session,
-		}).Fields, nil
+		}).Fields, 0, nil
 	}
 	return c.fallbackClient.Prepare(ctx, session, sql)
 }