--- a/go/mysql/fakesqldb/server.go
+++ b/go/mysql/fakesqldb/server.go
@@ -525,8 +525,8 @@ func (db *DB) comQueryOrdered(query string) (*sqltypes.Result, error) {
 }
 
 // ComPrepare is part of the mysql.Handler interface.
-func (db *DB) ComPrepare(*mysql.Conn, string) ([]*querypb.Field, error) {
-	return nil, nil
+func (db *DB) ComPrepare(*mysql.Conn, string) ([]*querypb.Field, uint16, error) {
+	return nil, 0, nil
 }
 
 // ComStmtExecute is part of the mysql.Handler interface.