--- a/go/mysql/server.go
+++ b/go/mysql/server.go
@@ -109,7 +109,7 @@ type Handler interface {
 
 	// ComPrepare is called when a connection receives a prepared
 	// statement query.
-	ComPrepare(c *Conn, query string) ([]*querypb.Field, error)
+	ComPrepare(c *Conn, query string) ([]*querypb.Field, uint16, error)
 
 	// ComStmtExecute is called when a connection receives a statement
 	// execute query.