--- a/go/vt/proto/vtgate/vtgate_vtproto.pb.go
+++ b/go/vt/proto/vtgate/vtgate_vtproto.pb.go
@@ -431,6 +431,7 @@ func (m *PrepareResponse) CloneVT() *PrepareResponse {
 	r := new(PrepareResponse)
 	r.Error = m.Error.CloneVT()
 	r.Session = m.Session.CloneVT()
+	r.ParamsCount = m.ParamsCount
 	if rhs := m.Fields; rhs != nil {
 		tmpContainer := make([]*query.Field, len(rhs))
 		for k, v := range rhs {
@@ -1760,6 +1761,11 @@ func (m *PrepareResponse) MarshalToSizedBufferVT(dAtA []byte) (int, error) {
 		i -= len(m.unknownFields)
 		copy(dAtA[i:], m.unknownFields)
 	}
+	if m.ParamsCount != 0 {
+		i = protohelpers.EncodeVarint(dAtA, i, uint64(m.ParamsCount))
+		i--
+		dAtA[i] = 0x20
+	}
 	if len(m.Fields) > 0 {
 		for iNdEx := len(m.Fields) - 1; iNdEx >= 0; iNdEx-- {
 			size, err := m.Fields[iNdEx].MarshalToSizedBufferVT(dAtA[:i])
@@ -2381,6 +2387,9 @@ func (m *PrepareResponse) SizeVT() (n int) {
 			l = e.SizeVT()
 			n += 1 + l + protohelpers.SizeOfVarint(uint64(l))
 		}
+	}
+	if m.ParamsCount != 0 {
+		n += 1 + protohelpers.SizeOfVarint(uint64(m.ParamsCount))
 	}
 	n += len(m.unknownFields)
 	return n
@@ -5905,6 +5914,25 @@ func (m *PrepareResponse) UnmarshalVT(dAtA []byte) error {
 				return err
 			}
 			iNdEx = postIndex
+		case 4:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ParamsCount", wireType)
+			}
+			m.ParamsCount = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return protohelpers.ErrIntOverflow
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.ParamsCount |= uint32(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
 		default:
 			iNdEx = preIndex
 			skippy, err := protohelpers.Skip(dAtA[iNdEx:])