--- a/go/vt/proto/vtgate/vtgate_vtproto.pb.go
+++ b/go/vt/proto/vtgate/vtgate_vtproto.pb.go
@@ -186,6 +186,7 @@ func (m *ExecuteRequest) CloneVT() *ExecuteRequest {
 	r.CallerId = m.CallerId.CloneVT()
 	r.Session = m.Session.CloneVT()
 	r.Query = m.Query.CloneVT()
+	r.Prepared = m.Prepared
 	if len(m.unknownFields) > 0 {
 		r.unknownFields = make([]byte, len(m.unknownFields))
 		copy(r.unknownFields, m.unknownFields)
@@ -1020,6 +1021,16 @@ func (m *ExecuteRequest) MarshalToSizedBufferVT(dAtA []byte) (int, error) {
 	if m.unknownFields != nil {
 		i -= len(m.unknownFields)
 		copy(dAtA[i:], m.unknownFields)
+	}
+	if m.Prepared {
+		i--
+		if m.Prepared {
+			dAtA[i] = 1
+		} else {
+			dAtA[i] = 0
+		}
+		i--
+		dAtA[i] = 0x40
 	}
 	if m.Query != nil {
 		size, err := m.Query.MarshalToSizedBufferVT(dAtA[:i])
@@ -2121,6 +2132,9 @@ func (m *ExecuteRequest) SizeVT() (n int) {
 	if m.Query != nil {
 		l = m.Query.SizeVT()
 		n += 1 + l + protohelpers.SizeOfVarint(uint64(l))
+	}
+	if m.Prepared {
+		n += 2
 	}
 	n += len(m.unknownFields)
 	return n
@@ -4128,6 +4142,26 @@ func (m *ExecuteRequest) UnmarshalVT(dAtA []byte) error {
 				return err
 			}
 			iNdEx = postIndex
+		case 8:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Prepared", wireType)
+			}
+			var v int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return protohelpers.ErrIntOverflow
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				v |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			m.Prepared = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := protohelpers.Skip(dAtA[iNdEx:])