… go/mysql/conn_test.go
… go/mysql/query_test.go
… go/mysql/server_test.go
… go/test/endtoend/cluster/cluster_process.go
… go/test/endtoend/cluster/cluster_util.go
… go/test/endtoend/messaging/message_test.go
… go/test/endtoend/preparestmt/benchmark_test.go
… go/test/endtoend/recovery/recovery_util.go
… go/test/endtoend/transaction/twopc/metric/metric_test.go
… go/test/endtoend/transaction/twopc/twopc_test.go
… go/test/endtoend/vtcombo/recreate/recreate_test.go
… go/test/endtoend/vtcombo/vttest_sample_test.go
… go/test/endtoend/vtgate/grpc_api/acl_test.go
… go/test/endtoend/vtgate/grpc_api/execute_test.go
… go/test/endtoend/vtgate/queries/reference/main_test.go
… go/test/utils/diff.go
… go/test/vschemawrapper/vschema_wrapper.go
… go/vt/key/destination.go
… go/vt/key/destination_test.go
… go/vt/proto/vtgate/vtgate.pb.go
… go/vt/sqlparser/normalizer_test.go
… go/vt/sqlparser/sql.go
… go/vt/sqlparser/sql.y
… go/vt/sqlparser/utils.go
… go/vt/srvtopo/resolver.go
… go/vt/srvtopo/resolver_test.go
… go/vt/sysvars/sysvars.go
… go/vt/topo/topoproto/destination.go
… go/vt/topo/topoproto/destination_test.go
… go/vt/vitessdriver/driver.go