## Description
This PR changes the way Vitess handles prepared statements by modifying the plan cache lookup mechanism. Instead of normalizing and parsing the SQL string before looking up a cached execution plan, we now use the raw SQL string as provided in the PREPARE statement. This eliminates unnecessary parsing, improves lookup efficiency, and reduces overhead in query execution.

## Problem Statement

Currently, when a client issues a PREPARE statement, vtgate:
1.	Parses the SQL query.
2.	Normalizes it to a canonical form.
3.	Uses the normalized query string to check for an existing cached execution plan.

## What to do About it

Instead of parsing and normalizing the query before checking the plan cache, we will:
	•	Directly store the original SQL string from the PREPARE statement in the cache.
	•	Use this unmodified SQL string for cache lookups.
	•	Only parse the query when the plan is not found in the cache.


## Benchmark

```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/test/endtoend/preparestmt
cpu: Apple M1 Max
                                        │ pfp-main.txt │            pfp-new.txt            │
                                        │    sec/op    │   sec/op     vs base              │
PreparedStmt/Insert-4                      494.3µ ± 3%   482.4µ ± 1%  -2.41% (p=0.002 n=6)
PreparedStmt/Select-4                      301.6µ ± 2%   288.4µ ± 3%  -4.35% (p=0.004 n=6)
PreparedStmt/Join_Select:Simple_Route-4    292.8µ ± 1%   269.9µ ± 1%  -7.81% (p=0.002 n=6)
PreparedStmt/Update-4                      344.9µ ± 2%   336.1µ ± 2%  -2.54% (p=0.009 n=6)
PreparedStmt/Delete-4                      311.4µ ± 0%   302.6µ ± 1%  -2.80% (p=0.002 n=6)
geomean                                    342.1µ        328.4µ       -4.01%

```

## Related Issue(s)

- Closes https://github.com/vitessio/vitess/issues/17621

## Checklist

-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required